+++ docs/conf.py
@@ -18,7 +18,11 @@
 #sys.path.insert(0, os.path.abspath('.'))
 
 # -- General configuration ------------------------------------------------
-
+nitpicky = True
+linkcheck_anchors_ignore = [".*"]
+linkcheck_ignore = [r'http://localhost:\d+/', r'https://github.com/mit-nlp/MITIE/releases/download/']
+linkcheck_retries = 2
+linkcheck_timeout = 5
 # If your documentation needs a minimal Sphinx version, state it here.
 #needs_sphinx = '1.0'