+++ rasa_nlu/model.py
@@ -1,23 +1,27 @@
-from __future__ import unicode_literals
-from __future__ import print_function
-from __future__ import division
 from __future__ import absolute_import
-from builtins import str
-from builtins import object
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
 import datetime
+import io
 import json
 import logging
 import os
-import io
+import copy
 
+from builtins import object
+from builtins import str
 from typing import Any
 from typing import Dict
 from typing import List
 from typing import Optional
 from typing import Text
 
-import rasa_nlu.components
+import rasa_nlu
+from rasa_nlu import components
 from rasa_nlu.components import Component
+from rasa_nlu.components import ComponentBuilder
 from rasa_nlu.config import RasaNLUConfig
 from rasa_nlu.persistor import Persistor
 from rasa_nlu.training_data import TrainingData
@@ -33,6 +37,9 @@ class InvalidModelError(Exception):
     def __init__(self, message):
         self.message = message
 
+    def __str__(self):
+        return self.message
+
 
 class Metadata(object):
     """Captures all necessary information about a model to load it and prepare it for usage."""
@@ -41,10 +48,12 @@ class Metadata(object):
     def load(model_dir):
         # type: (Text) -> 'Metadata'
         """Loads the metadata from a models directory."""
-
-        with io.open(os.path.join(model_dir, 'metadata.json'), encoding="utf-8") as f:
-            data = json.loads(f.read())
-        return Metadata(data, model_dir)
+        try:
+            with io.open(os.path.join(model_dir, 'metadata.json'), encoding="utf-8") as f:
+                data = json.loads(f.read())
+            return Metadata(data, model_dir)
+        except Exception as e:
+            raise InvalidModelError("Failed to load model metadata. {}".format(e))
 
     def __init__(self, metadata, model_dir):
         # type: (Dict[Text, Any], Optional[Text]) -> None
@@ -99,80 +108,41 @@ class Trainer(object):
     # Officially supported languages (others might be used, but might fail)
     SUPPORTED_LANGUAGES = ["de", "en"]
 
-    def __init__(self, config, component_builder=None):
-        # type: (RasaNLUConfig, Optional[rasa_nlu.components.ComponentBuilder]) -> None
+    def __init__(self, config, component_builder=None, skip_validation=False):
+        # type: (RasaNLUConfig, Optional[ComponentBuilder], bool) -> None
 
         self.config = config
+        self.skip_validation = skip_validation
         self.training_data = None
         self.pipeline = []
         if component_builder is None:
             # If no builder is passed, every interpreter creation will result in a new builder.
             # hence, no components are reused.
-            component_builder = rasa_nlu.components.ComponentBuilder()
+            component_builder = components.ComponentBuilder()
+
+        # Before instantiating the component classes, lets check if all required packages are available
+        if not self.skip_validation:
+            components.validate_requirements(config.pipeline)
 
         # Transform the passed names of the pipeline components into classes
         for component_name in config.pipeline:
             component = component_builder.create_component(component_name, config)
             self.pipeline.append(component)
 
-    def validate(self, allow_empty_pipeline=False):
-        # type: (bool) -> None
-        """Validates a pipeline before it is run. Ensures, that all arguments are present to train the pipeline."""
-
-        # Ensure the pipeline is not empty
-        if not allow_empty_pipeline and len(self.pipeline) == 0:
-            raise ValueError("Can not train an empty pipeline. " +
-                             "Make sure to specify a proper pipeline in the configuration using the `pipeline` key." +
-                             "The `backend` configuration key is NOT supported anymore.")
-
-        # Validate the init phase
-        context = {}
-
-        for component in self.pipeline:
-            try:
-                rasa_nlu.components.fill_args(component.pipeline_init_args(), context, self.config.as_dict())
-                updates = component.context_provides.get("pipeline_init", [])
-                for u in updates:
-                    context[u] = None
-            except rasa_nlu.components.MissingArgumentError as e:
-                raise Exception("Failed to validate at component '{}'. {}".format(component.name, e.message))
-
-        after_init_context = context.copy()
-
-        context["training_data"] = None     # Prepare context for testing the training phase
-
-        for component in self.pipeline:
-            try:
-                rasa_nlu.components.fill_args(component.train_args(), context, self.config.as_dict())
-                updates = component.context_provides.get("train", [])
-                for u in updates:
-                    context[u] = None
-            except rasa_nlu.components.MissingArgumentError as e:
-                raise Exception("Failed to validate at component '{}'. {}".format(component.name, e.message))
-
-        # Reset context to test processing phase and prepare for training phase
-        context = after_init_context
-        context["text"] = None
-
-        for component in self.pipeline:
-            try:
-                rasa_nlu.components.fill_args(component.process_args(), context, self.config.as_dict())
-                updates = component.context_provides.get("process", [])
-                for u in updates:
-                    context[u] = None
-            except rasa_nlu.components.MissingArgumentError as e:
-                raise Exception("Failed to validate at component '{}'. {}".format(component.name, e.message))
-
     def train(self, data):
         # type: (TrainingData) -> Interpreter
         """Trains the underlying pipeline by using the provided training data."""
+
+        # Before training the component classes, lets check if all arguments are provided
+        if not self.skip_validation:
+            components.validate_arguments(self.pipeline, self.config)
 
         self.training_data = data
 
         context = {}
 
         for component in self.pipeline:
-            args = rasa_nlu.components.fill_args(component.pipeline_init_args(), context, self.config.as_dict())
+            args = components.fill_args(component.pipeline_init_args(), context, self.config.as_dict())
             updates = component.pipeline_init(*args)
             if updates:
                 context.update(updates)
@@ -182,14 +152,16 @@ class Trainer(object):
         context["training_data"] = data
 
         for component in self.pipeline:
-            args = rasa_nlu.components.fill_args(component.train_args(), context, self.config.as_dict())
+            args = components.fill_args(component.train_args(), context, self.config.as_dict())
+            logging.info("Starting to train component {}".format(component.name))
             updates = component.train(*args)
+            logging.info("Finished training component.")
             if updates:
                 context.update(updates)
 
         return Interpreter(self.pipeline, context=init_context, config=self.config.as_dict())
 
-    def persist(self, path, persistor=None, create_unique_subfolder=True):
+    def persist(self, path, persistor=None, model_name=None):
         # type: (Text, Optional[Persistor], bool) -> Text
         """Persist all components of the pipeline to the passed path. Returns the directory of the persited model."""
 
@@ -199,12 +171,12 @@ class Trainer(object):
             "pipeline": [component.name for component in self.pipeline],
         }
 
-        if create_unique_subfolder:
+        if model_name is None:
             dir_name = os.path.join(path, "model_" + timestamp)
-            os.makedirs(dir_name)
         else:
-            dir_name = path
+            dir_name = os.path.join(path, model_name)
 
+        os.makedirs(dir_name)
         metadata.update(self.training_data.persist(dir_name))
 
         for component in self.pipeline:
@@ -215,7 +187,7 @@ class Trainer(object):
         Metadata(metadata, dir_name).persist(dir_name)
 
         if persistor is not None:
-            persistor.send_tar_to_s3(dir_name)
+            persistor.save_tar(dir_name)
         logging.info("Successfully saved model into '{}'".format(os.path.abspath(dir_name)))
         return dir_name
 
@@ -224,33 +196,41 @@ class Interpreter(object):
     """Use a trained pipeline of components to parse text messages"""
 
     # Defines all attributes (and their default values) that will be returned by `parse`
-    default_output_attributes = {"intent": {"name": "", "confidence": 0.0}, "entities": [], "text": ""}
+    @staticmethod
+    def default_output_attributes():
+        return {"intent": {"name": "", "confidence": 0.0}, "entities": [], "text": ""}
 
     @staticmethod
-    def load(meta, config, component_builder=None):
-        # type: (Metadata, RasaNLUConfig, Optional[rasa_nlu.components.ComponentBuilder]) -> Interpreter
+    def load(meta, config, component_builder=None, skip_valdation=False):
+        # type: (Metadata, RasaNLUConfig, Optional[ComponentBuilder], bool) -> Interpreter
         """Load a stored model and its components defined by the provided metadata."""
-        context = {"model_dir": meta.model_dir}
+        context = Interpreter.default_output_attributes()
+        context.update({"model_dir": meta.model_dir})
+
         if component_builder is None:
             # If no builder is passed, every interpreter creation will result in a new builder.
             # hence, no components are reused.
-            component_builder = rasa_nlu.components.ComponentBuilder()
+            component_builder = components.ComponentBuilder()
 
         model_config = config.as_dict()
         model_config.update(meta.metadata)
 
         pipeline = []
 
+        # Before instantiating the component classes, lets check if all required packages are available
+        if not skip_valdation:
+            components.validate_requirements(meta.pipeline)
+
         for component_name in meta.pipeline:
             component = component_builder.load_component(component_name, context, model_config, meta)
             try:
-                args = rasa_nlu.components.fill_args(component.pipeline_init_args(), context, model_config)
+                args = components.fill_args(component.pipeline_init_args(), context, model_config)
                 updates = component.pipeline_init(*args)
                 if updates:
                     context.update(updates)
                 pipeline.append(component)
-            except rasa_nlu.components.MissingArgumentError as e:
-                raise Exception("Failed to initialize component '{}'. {}".format(component.name, e.message))
+            except components.MissingArgumentError as e:
+                raise Exception("Failed to initialize component '{}'. {}".format(component.name, e))
 
         return Interpreter(pipeline, context, model_config)
 
@@ -258,7 +238,7 @@ class Interpreter(object):
         # type: (List[Component], Dict[Text, Any], Dict[Text, Any], Optional[Metadata]) -> None
 
         self.pipeline = pipeline
-        self.context = context
+        self.context = context if context is not None else {}
         self.config = config
         self.meta = meta
         self.output_attributes = [output for component in pipeline for output in component.output_provides]
@@ -271,9 +251,10 @@ class Interpreter(object):
             # Not all components are able to handle empty strings. So we need to prevent that...
             # This default return will not contain all output attributes of all components,
             # but in the end, no one should pass an empty string in the first place.
-            return self.default_output_attributes.copy()
+            return self.default_output_attributes()
 
         current_context = self.context.copy()
+        current_context.update(self.default_output_attributes())
 
         current_context.update({
             "text": text,
@@ -281,15 +262,15 @@ class Interpreter(object):
 
         for component in self.pipeline:
             try:
-                args = rasa_nlu.components.fill_args(component.process_args(), current_context, self.config)
+                args = components.fill_args(component.process_args(), current_context, self.config)
                 updates = component.process(*args)
                 if updates:
                     current_context.update(updates)
-            except rasa_nlu.components.MissingArgumentError as e:
-                raise Exception("Failed to parse at component '{}'. {}".format(component.name, e.message))
+            except components.MissingArgumentError as e:
+                raise Exception("Failed to parse at component '{}'. {}".format(component.name, e))
 
-        result = self.default_output_attributes.copy()
-        all_attributes = list(self.default_output_attributes.keys()) + self.output_attributes
+        result = self.default_output_attributes()
+        all_attributes = list(self.default_output_attributes().keys()) + self.output_attributes
         # Ensure only keys of `all_attributes` are present and no other keys are returned
         result.update({key: current_context[key] for key in all_attributes if key in current_context})
         return result