+++ rasa_nlu/train.py
@@ -12,6 +12,7 @@ from typing import Tuple
 
 from rasa_nlu.components import ComponentBuilder
 from rasa_nlu.converters import load_data
+from rasa_nlu.model import Interpreter
 from rasa_nlu.model import Trainer
 
 from rasa_nlu.config import RasaNLUConfig
@@ -31,7 +32,7 @@ def create_argparser():
     parser.add_argument('-d', '--data', default=None, help="file containing training data")
     parser.add_argument('-c', '--config', required=True, help="config file")
     parser.add_argument('-l', '--language', default=None, choices=['de', 'en'], help="model and data language")
-    parser.add_argument('-t', '--num_threads', default=1, type=int,
+    parser.add_argument('-t', '--num_threads', default=None, type=int,
                         help="number of threads to use during model training")
     parser.add_argument('-m', '--mitie_file', default=None,
                         help='file with mitie total_word_feature_extractor')
@@ -44,8 +45,8 @@ def create_persistor(config):
 
     persistor = None
     if "bucket_name" in config:
-        from rasa_nlu.persistor import Persistor
-        persistor = Persistor(config['path'], config['aws_region'], config['bucket_name'])
+        from rasa_nlu.persistor import get_persistor
+        persistor = get_persistor(config)
 
     return persistor
 
@@ -61,16 +62,17 @@ def init():
 
 
 def do_train(config, component_builder=None):
-    # type: (RasaNLUConfig, Optional[ComponentBuilder]) -> Tuple[Trainer, Text]
+    # type: (RasaNLUConfig, Optional[ComponentBuilder]) -> Tuple[Trainer, Interpreter, Text]
     """Loads the trainer and the data and runs the training of the specified model."""
 
+    # Ensure we are training a model that we can save in the end
+    # WARN: there is still a race condition if a model with the same name is trained in another subprocess
     trainer = Trainer(config, component_builder)
     persistor = create_persistor(config)
     training_data = load_data(config['data'])
-    trainer.validate()
-    trainer.train(training_data)
-    persisted_path = trainer.persist(config['path'], persistor)
-    return trainer, persisted_path
+    interpreter = trainer.train(training_data)
+    persisted_path = trainer.persist(config['path'], persistor, model_name=config['name'])
+    return trainer, interpreter, persisted_path
 
 
 if __name__ == '__main__':