+++ CHANGELOG.rst
@@ -11,13 +11,44 @@ This project adheres to `Semantic Versioning`_ starting with version 0.7.0.
 
 Added
 -----
+
+Changed
+-------
+
+Removed
+-------
+
+Fixed
+-----
+
+[0.8.2] - 2017-05-08
+^^^^^^^^^^^^^^^^^^^^
+
+Fixed
+-----
+- deepcopy of context #343
+
+[0.8.1] - 2017-05-08
+^^^^^^^^^^^^^^^^^^^^
+
+Fixed
+-----
+- NER training reuses context inbetween requests
+
+[0.8.0] - 2017-05-08
+^^^^^^^^^^^^^^^^^^^^
+Added
+-----
 - ngram character featurizer (allows better handling of out-of-vocab words)
 - replaced pre-wired backends with more flexible pipeline definitions
 - return top 10 intents with sklearn classifier `#199 <https://github.com/golastmile/rasa_nlu/pull/199>`_
 - python type annotations for nearly all public functions
-- support for arbitrary spacy model names
+- support for arbitrary spacy language model names
 - duckling components to provide normalized output for structured entities
 - Conditional random field entity extraction (Markov model for entity tagging, better named entity recognition with low and medium data and similarly well at big data level)
+- allow naming of trained models instead of generated model names
+- dynamic check of requirements for the different components & error messages on missing dependencies
+- support for using multiple entity extractors and combining results downstream
 
 Changed
 -------
@@ -26,16 +57,22 @@ Changed
 - when loading data in a foreign format (api.ai, luis, wit) the data gets properly split into intent & entity examples
 - Configuration:
     - added ``max_number_of_ngrams``
-    - added ``pipeline``
+    - removed ``backend`` and added ``pipeline`` as a replacement
     - added ``luis_data_tokenizer``
-    - removed ``backend``
+    - added ``duckling_dimensions``
 - parser output format changed
     from ``{"intent": "greeting", "confidence": 0.9, "entities": []}``
 
     to ``{"intent": {"name": "greeting", "confidence": 0.9}, "entities": []}``
+- entities output format changed
+    from ``{"start": 15, "end": 28, "value": "New York City", "entity": "GPE"}``
+
+    to ``{"extractor": "ner_mitie", "processors": ["ner_synonyms"], "start": 15, "end": 28, "value": "New York City", "entity": "GPE"}``
+
+    where ``extractor`` denotes the entity extractor that originally found an entity, and ``processor`` denotes components that alter entities, such as the synonym component.
 - camel cased MITIE classes (e.g. ``MITIETokenizer`` → ``MitieTokenizer``)
 - model metadata changed, see migration guide
-- updated to spacy 1.7 (breaks existing spacy models!)
+- updated to spacy 1.7 and dropped training and loading capabilities for the spacy component (breaks existing spacy models!)
 - introduced compatibility with both Python 2 and 3
 
 Removed
@@ -47,6 +84,14 @@ Fixed
 - support entity only training `#181 <https://github.com/golastmile/rasa_nlu/issues/181>`_
 - resolved conflicts between metadata and configuration values `#219 <https://github.com/golastmile/rasa_nlu/issues/219>`_
 - removed tokenization when reading Luis.ai data (they changed their format) `#241 <https://github.com/golastmile/rasa_nlu/issues/241>`_
+
+[0.7.4] - 2017-03-27
+^^^^^^^^^^^^^^^^^^^^
+
+Fixed
+-----
+- fixed failed loading of example data after renaming attributes, i.e. "KeyError: 'entities'"
+
 [0.7.3] - 2017-03-15
 ^^^^^^^^^^^^^^^^^^^^